establishing assignment baseline as branch lab1 before moving on to assignment branch lab2
