+++ .eslintrc
@@ -1,0 +1,28 @@
+{
+    "env": {
+        "es6": true,
+        "node": true,
+        "mocha": true,
+        "regex": true
+    },
+    "extends": "eslint:recommended",
+    "rules": {
+        "no-console": 0,
+        "indent": [
+            "error",
+            2
+        ],
+        "linebreak-style": [
+            "error",
+            "unix"
+        ],
+        "quotes": [
+            "error",
+            "single"
+        ],
+        "semi": [
+            "error",
+            "always"
+        ]
+    }
+}
