+++ index.js
@@ -1,0 +1,5 @@
+// establish 'tcpServer.js' as starter button for tcp server app
+
+var tcpServer = require('./tcpServer');
+
+console.log(tcpServer('class'));
