+++ tcpServer.js
@@ -21,4 +21,4 @@ const port = 65000;
 server.listen(port, err => {
   if (err) console.log('ERROR!', err);
   else console.log('server listening on port', port);
-})
+});
